+++ src/droplet/create-droplet/create-droplet.ts
@@ -12,7 +12,7 @@ export interface ICreateDropletApiRequest {
   monitoring?: boolean;
   name: string;
   private_networking?: boolean;
-  region: string;
+  region?: string;
   size: string;
   ssh_keys?: Array<string | number>;
   tags?: string[];